requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pandas==2.1.3
openpyxl==3.1.2
phonenumbers==8.13.26
email-validator==2.1.0.post1
pycountry==22.3.5
PyJWT==2.8.0
httpx==0.25.1
python-multipart==0.0.6
alembic==1.12.1
pyyaml==6.0.1
psycopg2-binary==2.9.9
jinja2==3.1.6
aiosmtplib==5.1.0
